refactor(models): transmit http models to the corresponding place
