+++ internal/model/route.go
@@ -3,5 +3,5 @@ package model
 type Route struct {
 	BusID  int64 `db:"bus_id"`
 	StopID int64 `db:"stop_id"`
-	Step   uint8 `db:"step"`
+	Step   int8  `db:"step"`
 }
