--- internal/model/v1/bus.go
@@ -1,7 +1,0 @@
-package v1
-
-type Bus struct {
-	ID   int64  `json:"id,omitempty"`
-	City string `json:"city"`
-	Num  string `json:"num"`
-}
