--- internal/model/v1/city.go
@@ -1,6 +1,0 @@
-package v1
-
-type City struct {
-	ID   int    `json:"id,omitempty"`
-	Name string `json:"name"`
-}
