--- internal/model/v1/route.go
@@ -1,7 +1,0 @@
-package v1
-
-type Route struct {
-	BusID  int64 `json:"bus_id"`
-	StopID int64 `json:"stop_id"`
-	Step   uint8 `json:"step"`
-}
